% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_text.R
\name{convert_text}
\alias{convert_text}
\title{Convert Text}
\usage{
convert_text(label)
}
\arguments{
\item{label}{A string to be processed}
}
\value{
A string with unicode in place of html markdown
}
\description{
do a simple swap of html based characters for special symbols needed in titles & labels
}
\examples{
title <- "Is it finally &Pi; day?"
title <- convert_text(title)
}
