% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{env_var_exists}
\alias{env_var_exists}
\title{Check for Env & Var}
\usage{
env_var_exists(envName, varName = NULL, msg = TRUE)
}
\arguments{
\item{envName}{text of the environment}

\item{varName}{text of the variable name}

\item{msg}{T/F should a message be printed?}
}
\value{
T/F if both exists
}
\description{
Simple function to do the routine checking of if an environment & variable
exists. T/F is the return, a warning message is also thrown if msg is TRUE (default).
Msg can be set to FALSE when checking to ensure env or var doesn't exist.
If var is left at NULL it will only check for environment.
}
\examples{
env_var_exists("the", "envList")
}
