% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default.R
\name{get_default}
\alias{get_default}
\title{Get Default Value}
\usage{
get_default(envName, varName)
}
\arguments{
\item{envName}{text of the environment}

\item{varName}{text of the variable}
}
\value{
value from env, NULL if not found
}
\description{
Submit an environment and variable name and get returned the default value.
NULL response indicates unable to find environment and/or variable.
}
\examples{
get_default('the', 'envList')
}
