% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_suffix}
\alias{get_suffix}
\title{Return file suffix}
\usage{
get_suffix(location.file)
}
\arguments{
\item{location.file}{text string of the filename}
}
\value{
text string, NULL if not found or too short
}
\description{
Return the text following the final '.' in the name. If
}
\examples{
get_suffix("file.txt")
}
