% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histova_example.R
\name{histova_example}
\alias{histova_example}
\title{Read Example}
\usage{
histova_example(path = NULL)
}
\arguments{
\item{path}{The name of the file being searched for}
}
\value{
files associated with the package if path is NULL, the file location for a specific file if defined or error
}
\description{
Read Example
}
