% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histova_msg.R
\name{histova_msg}
\alias{histova_msg}
\title{Histova Message}
\usage{
histova_msg(
  msg,
  type = "msg",
  breaker = FALSE,
  tabs = 0,
  PRINT = TRUE,
  LOG = TRUE,
  hsa = NA
)
}
\arguments{
\item{msg}{Simple text string to print out / write to file}

\item{type}{title/head/subhead/msg/warn/verbose For what type of print call to use (title, head & subhead include formatting)}

\item{breaker}{above/below/both Should the text string have border lines before and/or after it}

\item{tabs}{Any indentation for the message (only compatible with msg and warn)}

\item{PRINT}{T/F Print out to screen}

\item{LOG}{T/F Write to .histova log file}

\item{hsa}{R6 histova object for settings details}
}
\description{
Histova Message
}
