% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_color}
\alias{is_color}
\title{Is Valid Color?}
\usage{
is_color(color)
}
\arguments{
\item{color}{variable to be checked if it is a R color}
}
\value{
TRUE if color, FALSE if not
}
\description{
Send in a variable to see if it is an acceptable color with R, checks against
various notations including transparant values
}
