% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_con_open}
\alias{is_con_open}
\title{Is Connection Open}
\usage{
is_con_open(con)
}
\arguments{
\item{con}{variable to be checked if it is a file connection}
}
\value{
TRUE if open, FALSE if not open or not connection
}
\description{
Send in a variable that might be a connection to check... this simply catches
the error thrown by base::isOpen when a variable is sent in that has already
been closed...
}
