% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_num}
\alias{is_num}
\title{Is Numeric}
\usage{
is_num(num)
}
\arguments{
\item{num}{variable to be checked if it is a number}
}
\value{
TRUE if numeric, FALSE if not
}
\description{
Is the value a legit number or not, this includes decimals and negatives
}
