% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load Data}
\usage{
load_data()
}
\description{
Load the data in from the file and perform any global manipulations. Following the file header
(all lines that begin with '#') comes the data for the figure. The format is
value\if{html}{\out{<tab>}}Group1\if{html}{\out{<tab>}}Group2
Group 2 is optional. The data is displayed in the order it is entered into the file.
All data is saved to the histova environment for use by other functions.
}
\examples{
load_data()
}
