% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_file.R
\name{load_file}
\alias{load_file}
\title{Load File}
\usage{
load_file(location.dir, location.file, savePlot)
}
\arguments{
\item{location.dir}{The directory the data file is contained in}

\item{location.file}{The file containing the data}

\item{savePlot}{Should the finished plot & log be saved to disk}
}
\description{
This function loads the settings and data contained in the specified file
into the environment. A copy of the data in the file should be saved
within a 'raw' variable. The settings can be changed with a backup
assumed to be in the initial file. The goal is that once the aesthetic settings
are finalized the settings and raw data can be written back out to a file
for subsequent re-use.
}
\examples{
load_file("/Users/Shared/HISTOVA_DATA", "test.txt", savePlot = FALSE)
}
