% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_set.R
\name{opt_set}
\alias{opt_set}
\title{Set an option / setting}
\usage{
opt_set(envName, varName, varData)
}
\arguments{
\item{envName}{name of the environment where the option is located}

\item{varName}{name of the variable to edit}

\item{varData}{data to set the variable to}
}
\value{
Pass/Fail message
}
\description{
Function simply sets an option within the package. Run 'opt_print()' to get an overview of the
existing options. Mainly designed to allow for quick changes to the aesthetic settings for the
figure.
}
\examples{
opt_set("fig", "X.Angle", 55)
}
