% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_env_dict.R
\name{print_env_dict}
\alias{print_env_dict}
\title{Print Environment Dictionary}
\usage{
print_env_dict(env, name, rootEnv, depth = 0, includeData = TRUE)
}
\arguments{
\item{env}{the environment object to traverse}

\item{name}{name of the environment object for human printing}

\item{rootEnv}{the environment that actually contains the data}

\item{depth}{in num of tabs for printing out dictionary tree}

\item{includeData}{T/F should data be printed out alongside name}
}
\value{
depth of print in tabs
}
\description{
Simple and largely internal function that traverses a pseudo dictionary composed of environment variables
to dynamically print out variable information that this package depends on. Function is designed to be a
guide for editing or loading settings for figure design to allow for customization of figure
while loaded in R.
}
