% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_anova.R
\name{run_anova}
\alias{run_anova}
\title{Run ANOVA}
\usage{
run_anova()
}
\description{
Calculate the ANOVA values for the specified groups. Depending on the configuration the ANOVA
is either run within each Group2 or between all possible Group1_Group2 combinations. All of the
data is pulled from environment variables.
}
