% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_data.R
\name{run_data}
\alias{run_data}
\title{Run Data}
\usage{
run_data(hsa)
}
\arguments{
\item{hsa}{the histova R6 object}
}
\description{
This function is a wrapper to run through all of the statistical operations
typically carried out on a set of data. This includes any tests as well as
modifications and transformations of the data itself.
}
\details{
All relevant data and settings should be stored in the environment variables
that are accessed by this function.
}
