% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stats.R
\name{run_stats}
\alias{run_stats}
\title{Run requested stats}
\usage{
run_stats(hsa)
}
\arguments{
\item{hsa}{the histova R6 object}
}
\description{
Wrapper that kicks off any statistical analysis that have been requested in the file.
}
