% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stats_prep.R
\name{run_stats_prep}
\alias{run_stats_prep}
\title{Run Stats Prep}
\usage{
run_stats_prep()
}
\description{
Calculate the stats (outlier removal, anova, TUKEY post hoc, summary table) that histova needs.
This function accesses the raw data that has been loaded into the environment and then calls
the appropriate functions depending on what tests were specified in the config file...

No examples specified as the function \emph{depends} on having data loaded into the environment by other functions.
}
