% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sttest.R
\name{run_sttest}
\alias{run_sttest}
\title{Run STTest}
\usage{
run_sttest()
}
\description{
Run Students TTest against defined groups. The type of test and the groups to
be compared are assigned in the config file. The test is run off of data loaded
into the package's environment by previous functions.
}
