% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_transform.R
\name{run_transform}
\alias{run_transform}
\title{Run Transform}
\usage{
run_transform()
}
\description{
Carry out any desired transformations on the dataset. Currently the only transformation
available is 'treatment over control'. As with most functions in this package the
data controlling the behavior of the function is stored in the package environment
and the data being manipulated is also in the package environment.
}
