% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_env.R
\name{set_env}
\alias{set_env}
\title{Setup the environment}
\usage{
set_env(location.dir, location.file, saveLog, env.new = TRUE, printMsg = TRUE)
}
\arguments{
\item{location.dir}{The directory the data file is contained in}

\item{location.file}{The file containing the data}

\item{saveLog}{Should the log be saved to disk}

\item{env.new}{T/F is this a new environment or continued execution on an old one}

\item{printMsg}{T/F print a resume header message, only option when env.new is FALSE}
}
\value{
hsa R6 histova object created in this function
}
\description{
This function sets up the general environment, checking to make sure
the needed environments for data & options storage exist and
that the specified file exists
}
\details{
The log file is also initiated here.
}
\examples{
set_env("/Users/Shared/HISTOVA_DATA", "test.txt", saveLog = FALSE)
}
