% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_file.R
\name{write_file}
\alias{write_file}
\title{Write settings file}
\usage{
write_file(location.dir = NULL, location.file = NULL, overwrite = FALSE)
}
\arguments{
\item{location.dir}{directory to look for the file, NULL use environment}

\item{location.file}{name of the file, NULL use environment}

\item{overwrite}{T/F if there is an existing file rename, F simply overwrite}
}
\description{
Writes the functions currently loaded into the environment out to a file. If location.dir
or location.file are NULL it will attempt to pull them from the environment. If not present
an error message will be thrown. Overwrite determines if the function renames any existing
file at this location appending the current date&time (-YYYYMMDD_HHMMSS).
}
\examples{
write_file('dir', 'file')
}
