% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_figure.R
\name{build_figure}
\alias{build_figure}
\title{Build Figure}
\usage{
build_figure(hsa, printPlot = FALSE, savePlot = TRUE, printEnvMsg = TRUE)
}
\arguments{
\item{hsa}{the histova R6 object}

\item{printPlot}{T/F Should the finished plot be printed (def F)}

\item{savePlot}{T/F Should the finished plot & log be saved to disk (def T)}

\item{printEnvMsg}{T/F Passed on to set_env to determine if resume message printed (def T)}
}
\description{
This function uses the aesthetic settings as defined in the environment and the
data and constructs a corresponding figure with the contained information. The
only inputs are optional defining if the plot should be printed and / or if
it should be saved to disk.
}
