% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_figure.R
\name{build_figure}
\alias{build_figure}
\title{Build Figure}
\usage{
build_figure(printPlot = FALSE, savePlot = TRUE)
}
\arguments{
\item{printPlot}{Should the finished plot be printed}

\item{savePlot}{Should the finished plot & log be saved to disk}
}
\description{
This function uses the aesthetic settings as defined in the environment and the
data and constructs a corresponding figure with the contained information. The
only inputs are optional defining if the plot should be printed and / or if
it should be saved to disk.
}
