% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_histo.R
\name{build_histo}
\alias{build_histo}
\title{Build Histogram}
\usage{
build_histo(hsa)
}
\arguments{
\item{hsa}{active R6 histova object}
}
\description{
Build histogram using data and configuration settings saved in the package environment. This function
tries to put all the possible options specified in the configuration together into a working figure.
Package is typically called by generate_figure after environment variables have been loaded and all
desired statistical tests have been carried out.
}
