% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_histo.R
\name{build_histo}
\alias{build_histo}
\title{Build Histogram}
\usage{
build_histo()
}
\description{
\strong{CURRENT FUNCTIONING FIGURE!}\if{html}{\out{<br>}}
Basic histogram using above created summary and adding error bars!
}
