% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_figure.R
\name{generate_figure}
\alias{generate_figure}
\title{Generate Figure}
\usage{
generate_figure(
  location.dir,
  location.file,
  printPlot = FALSE,
  savePlot = TRUE
)
}
\arguments{
\item{location.dir}{The directory the data file is contained in}

\item{location.file}{The file containing the data}

\item{printPlot}{T/F Should the finished plot be printed (default to FALSE)}

\item{savePlot}{T/F Should the finished plot be saved to disk (default to TRUE)}
}
\description{
putting it all together
set the directory and file location first and then call this function to run through the specified file
}
\examples{
generate_figure("/Users/Shared/HISTOVA_DATA", "test.txt", savePlot = FALSE)

}
