% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_label_df.R
\name{generate_label_df}
\alias{generate_label_df}
\title{Generate Label df}
\usage{
generate_label_df(n)
}
\arguments{
\item{n}{The number of label subset (when Group2 is being compared separately)}
}
\value{
A data frame of labels
}
\description{
Generate labels for the figures. This is typically called from within build_histo(). As
is typical for this package most of the information needed to run is stored in the
package environment. However this function is not editing that many environment variables
and is returning a data frame.
}
