% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_label_df.R
\name{generate_label_df}
\alias{generate_label_df}
\title{Generate Label df}
\usage{
generate_label_df(n)
}
\arguments{
\item{n}{the number of labels}
}
\value{
A data frame of labels
}
\description{
generate labels for the figure
}
