% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histova.R
\name{histova}
\alias{histova}
\title{R6 Class Representing a HISTOVA object}
\description{
R6 Class Representing a HISTOVA object

R6 Class Representing a HISTOVA object
}
\details{
Store all of the data and provide a get() function
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{behave}}{list for object behavior, default to FALSE}

\item{\code{fig}}{list for figure data}

\item{\code{file}}{list for histova including open file connections}

\item{\code{notes}}{list for histova process}

\item{\code{raw}}{list for histova input data for fallback after transformation}

\item{\code{stats}}{list for stats output}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HISTOVA-get}{\code{histova$get()}}
\item \href{#method-HISTOVA-get_default}{\code{histova$get_default()}}
\item \href{#method-HISTOVA-is_style}{\code{histova$is_style()}}
\item \href{#method-HISTOVA-clone}{\code{histova$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HISTOVA-get"></a>}}
\if{latex}{\out{\hypertarget{method-HISTOVA-get}{}}}
\subsection{Method \code{get()}}{
get data from object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{histova$get(x, ref = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{data to pull}

\item{\code{ref}}{reference HISTOVA object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HISTOVA-get_default"></a>}}
\if{latex}{\out{\hypertarget{method-HISTOVA-get_default}{}}}
\subsection{Method \code{get_default()}}{
if there is a default value available pull it, otherwise return NULL
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{histova$get_default(name.L, name.I)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name.L}}{the list name}

\item{\code{name.I}}{the info name / variable}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HISTOVA-is_style"></a>}}
\if{latex}{\out{\hypertarget{method-HISTOVA-is_style}{}}}
\subsection{Method \code{is_style()}}{
is this a style variable that can be overidden?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{histova$is_style(name.L, name.I)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name.L}}{the list name}

\item{\code{name.I}}{the info name / variable}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HISTOVA-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HISTOVA-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{histova$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
