% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histova_msg.R
\name{histova_msg}
\alias{histova_msg}
\title{histova message}
\usage{
histova_msg(
  msg,
  type = "msg",
  breaker = NULL,
  tabs = 0,
  PRINT = TRUE,
  LOG = TRUE
)
}
\arguments{
\item{msg}{Simple text string to print out / write to file}

\item{type}{msg/warn For what type of print call to use}

\item{breaker}{Should the text string have border lines before and/or after it}

\item{tabs}{And indentation for the message}

\item{PRINT}{T/F Print out to screen}

\item{LOG}{T/F Write to .histova log file}
}
\description{
histova message
}
