% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_vars.R
\name{init_vars}
\alias{init_vars}
\title{Initialize variables to defaults}
\usage{
init_vars()
}
\description{
Reset the variables to the defaults used to load and generate another figure
}
\examples{
init_vars()
}
