% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_file_head.R
\name{load_file_head}
\alias{load_file_head}
\title{Load File Head}
\usage{
load_file_head(hsa)
}
\description{
Load the configuration from the head of the file specified in the$Location.file. Each
configuration file should begin with the settings section that details the aesthetics of
the desired figure as well as what kind of statistical tests to carry out. These settings
are all loaded into a few environments within the histova package and stored there
to be accessible between functions. The Override option is controlled in this function.
}
