% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_file_head.R
\name{load_file_head}
\alias{load_file_head}
\title{Load File Head}
\usage{
load_file_head()
}
\description{
Load the configuration from the head of the file specified in the$Location.file
}
\examples{
load_file_head()
}
