% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_outlier.R
\name{run_outlier}
\alias{run_outlier}
\title{Run Outlier}
\usage{
run_outlier()
}
\description{
remove any outliers, if requested; DEPENDS on already having loaded data to function as it operates on saved environment variables
}
