% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_aesthetics.R
\name{set_aesthetics}
\alias{set_aesthetics}
\title{Set Aesthetics}
\usage{
set_aesthetics()
}
\description{
Generate the lists and variables needed for displaying the correct aesthetics in the figure
current focus is on the color of the bars & the points being used along with size

output the correct set of named lists
\if{html}{\out{<br>}}Fig.Colors -> list for the histogram
\if{html}{\out{<br>}}Fig.Scatter.Colors.List -> colors for the scatter
\if{html}{\out{<br>}}Fig.Scatter.Shape.List -> list of shapes
\if{html}{\out{<br>}}Fig.Scatter.Size.List -> list of sizes

output of the stroke size
\if{html}{\out{<br>}}Fig.Scatter.Stroke -> the one size for stroking the scatter
}
